requirements.txt
asgiref==3.2.3
certifi==2019.11.28
chardet==3.0.4
defusedxml==0.6.0
dj-static==0.0.6
Django==3.0.3
django-allauth==0.41.0
django-appconf==1.0.3
django-compressor==2.4
django-cors-headers==3.2.1
django-forms-bootstrap==3.1.0
django-oauth-toolkit==1.2.0
djangorestframework==3.11.0
docker-py==1.10.6
docker-pycreds==0.4.0
dockerpty==0.4.1
docopt==0.6.2
gunicorn==20.0.4
idna==2.9
oauthlib==3.1.0
psycopg2-binary==2.8.4
python3-openid==3.1.0
pytz==2019.3
PyYAML==5.3
rcssmin==1.0.6
requests==2.23.0
requests-oauthlib==1.3.0
rjsmin==1.1.0
six==1.14.0
sqlparse==0.3.1
static3==0.7.0
texttable==1.6.2
urllib3==1.25.8
websocket-client==0.57.0

